requirements.txt
fastapi
uvicorn
inference-sdk
opencv-python
pandas
reportlab
matplotlib
seaborn
python-multipart
python-dotenv
sqlalchemy
psycopg[binary]
google-auth
python-jose[cryptography]